Add q_remove_head()
The q_remove_head() function enables removal of elements from the
head of a queue. It adheres to the "remove" concept by unlinking the
element without deallocating memory. And It copies the removed
element's value to a provided buffer (sp) with proper size constraints
and null termination. Finally, it returns the removed element.