Add q_swap()
Implement q_swap function that swap every two values of adjacent nodes.
The implementation leverages the XOR swap technique to avoid
unneccessary copies and memory allocations. In order to XOR swap values
of nodes, it performs a cast from char * to uintptr_t that is capable
to hold object pointers.