Implement preemptive scheduling
Introduced a preemptive coroutine scheduling mechanism to manage
concurrent execution of tasks in the AI vs. AI game mode. This approach
enhances the efficiency of the game by allowing tasks to yield control
to other tasks

Enabled a new game mode (Mode 2) that allows for 10 rounds of AI vs. AI
play. Implemented a timer_wait() mechanism to provide a brief pause
between AI moves, allowing users to observe the game's progression and
the AI's strategies.

Additionally, two functions have been added. One is to press "p" to
pause a round of the game. Another is the addition of the ability to
toggle the visibility of the game screen.